DC_MakeHelperWaves: Init OscilloscopeData DA channels with zero for power spectrum display
In 1f2ce19b (DC_InitDataHoldingWave: Initialize all channels to NaN,
2019-10-22) we switched to setting the initial values for all channels
to NaN.

This broke the power spectrum as

FFT/OUT=4/COLS ...

does not work as expected. The reason is that /OUT can not be combined
with /COLS.

Now we could drop the /OUT flag and call magsqr ourselves. But that
would be too slow, so let's initialize the DA channels to 0 in this
case.